+++ framework-crates/objc2-ui-kit/src/lib.rs
@@ -17,7 +17,6 @@ extern crate alloc;
 extern crate std;
 
 mod generated;
-mod geometry;
 #[cfg(feature = "UIGestureRecognizer")]
 mod gesture_recognizer;
 #[cfg(feature = "UIPasteConfigurationSupporting")]
@@ -31,13 +30,7 @@ mod text;
 
 #[allow(unused_imports, unreachable_pub)]
 pub use self::generated::*;
-pub use self::geometry::*;
 #[cfg(feature = "UIResponder")]
 pub use self::responder::*;
 #[cfg(feature = "NSText")]
 pub use self::text::*;
-
-// Used by UIKeyConstants
-// NOTE: CFIndex is c_long_long on __LLP64__ / Windows 64-bit (doesn't matter for us)
-#[allow(unused)]
-pub(crate) type CFIndex = core::ffi::c_long;